+++ Files/filedate/utils/__init__.py
@@ -1,3 +1,4 @@
 from .copy_file_date import *
 from .keep_file_date import *
 from .from_file_name import *
+from .from_metadata  import *